train: Return MLflow tracking information
Extend the API to return MLflow tracking information as part of a
training or evaluation response, including the experiment and run IDs,
and update the tests accordingly. If training is already in progress,
the API returns the experiment and run IDs of the current training run.
This affects the following routes:
* POST /train_supervised
* POST /train_unsupervised
* POST /train_unsupervised_with_hf_hub_dataset
* POST /train_metacat
* POST /evaluate

Signed-off-by: Phoevos Kalemkeris <phoevos.kalemkeris@ucl.ac.uk>