compose: Set top-level project name
Set the top-level project name in the docker-compose files so that it
can be used as the default prefix for container names. Use 'cms' for all
stacks apart from the load tests (i.e. 'cms-load') and the dev
deployment (i.e. 'dev-cms'). The project name can be overridden on
deployment through the '-p' command line flag or by setting the
'COMPOSE_PROJECT_NAME' environment variable.

Signed-off-by: Phoevos Kalemkeris <phoevos.kalemkeris@ucl.ac.uk>