fix: Wait for MinIO Server before creating bucket
Wait for the MinIO Server to be healthy before attempting to create a
bucket as part of the 'model-bucket-init' service. This is necessary
because the MinIO Server may not be ready to accept requests when the
service starts, resulting in errors like the following:

```
Unable to initialize new alias from the provided credentials. Server
not initialized, please try again.
```

Signed-off-by: Phoevos Kalemkeris <phoevos.kalemkeris@ucl.ac.uk>