fix: Chain bucket init commands to catch errors
Chain the commands involved in creating a MinIO bucket run as part of
the 'model-bucket-init' service in the 'docker-compose-mlflow.yml' file
to correctly catch errors and trigger service restarts. Previously, if
the first command (i.e. adding a MinIO host) failed, the second one
would still run, overwriting the error code and preventing the service
from restarting. Using the '&&' operator to chain the commands ensures
that the second command will only run if the first one is successful and
that the error code is correctly propagated, allowing Docker to retry in
the event of a failure. Note that the MinIO client commands can fail due
to a variety of reasons, such as network issues (some of which might be
transient), the MinIO Server not being ready, or incorrect credentials.

Signed-off-by: Phoevos Kalemkeris <phoevos.kalemkeris@ucl.ac.uk>