api: Include labels in DeID service model card
Extend the base model card schema to allow passing a mapping of labels,
i.e. concept IDs (e.g. CUIs) to names. This is particularly useful
when using the MedCAT model DeID service, which is typically configured
to serve models with a relatively small CDB that can be exposed in its
entirety as part of the model card. This information can be interesting
to the user (or client application), providing insight into the concepts
that the model is capable of identifying and by extension enabling them
to decide and control which of them should be redacted. The 'labels'
field is always included in the API's responses when the '/info'
endpoint is called, but is only populated with the mapping in a dict
format when the CMS instance is configured to use the MedCAT model DeID
service.

Signed-off-by: Phoevos Kalemkeris <phoevos.kalemkeris@ucl.ac.uk>