feat: Add MedCAT OPCS-4 service (#18)
Introduce a MedCAT OPCS-4 service to handle OPCS annotation codes,
corresponding to interventions and procedures. The new service mirrors
the structure and behaviour of the ICD-10 equivalent, exposing only the
OPCS-4 labels among the annotations generated by the underlying MedCAT
model, relying on the existence of a 'cui2opcs4' mapping in the latter's
concept database. This commit introduces the new model type and service,
adjusts the existing tests accordingly, extends the Docker Compose stack
with a 'medcat-opcs4' Docker service, and updates the Grafana/Prometheus
configuration to take it into account.

Signed-off-by: Phoevos Kalemkeris <phoevos.kalemkeris@ucl.ac.uk>